stat.st_size is apparently a signed value - so change code to avoid comparing signed and unsigned

+++ EventFilter/Utilities/plugins/RecoEventOutputModuleForFU.h
@@ -198,8 +198,8 @@ namespace evf {
       struct stat istat;
       stat(openDatFilePath_.string().c_str(), &istat);
       unsigned long readInput=0;
-      while (readInput<istat.st_size) {
-          unsigned long toRead=  readInput+1024*1024 < istat.st_size ? 1024*1024 : istat.st_size-readInput;
+      while (readInput<(unsigned int)istat.st_size) {
+	unsigned long toRead=  readInput+1024*1024 < (unsigned int)istat.st_size ? 1024*1024 : istat.st_size-readInput;
           fread(outBuf_,toRead,1,src);
           fwrite(outBuf_,toRead,1,des);
           readInput+=toRead;
